Pas2C: Fix bad C typedefs for Pascal data types Extended and Real

+++ project_files/hwc/rtl/pas2c.h
@@ -49,8 +49,8 @@ typedef int64_t Int64;
 typedef LongWord Cardinal;
 
 typedef LongInt Integer;
-typedef float extended;
-typedef float real;
+typedef long double extended;
+typedef double real;
 typedef float single;
 
 typedef bool boolean;
